+++ pkg/setup/setup.go
@@ -351,6 +351,12 @@ func (s *Setup) installDefaultPlugins(httpClient *httpclient.Client) error {
 // installPackageServicesPlugins installs CLI plugins for the services currently installed on the cluster.
 // When different versions of the same package are installed, it installs the plugin for the highest version.
 func (s *Setup) installPackageServicesPlugins(wg *sync.WaitGroup, httpClient *httpclient.Client, pbar *mpb.Progress) {
+	// Install all package CLIs when the env var is present.
+	installPackageCLIs, _ := s.envLookup("DCOS_CLI_EXPERIMENTAL_AUTOINSTALL_PACKAGE_CLIS")
+	if installPackageCLIs == "" {
+		return
+	}
+
 	cosmosClient, err := cosmos.NewClient()
 	if err != nil {
 		s.logger.Debug(err)
@@ -358,48 +364,33 @@ func (s *Setup) installPackageServicesPlugins(wg *sync.WaitGroup, httpClient *ht
 	}
 
 	pkgMap := make(map[string]dcosclient.CosmosPackage)
-
-	cErr := make(chan error, 1)
-	cResult := make(chan dcosclient.CosmosPackageListV1Response, 1)
-	go func() {
-		result, _, err := cosmosClient.PackageList(context.TODO(), &dcosclient.PackageListOpts{
-			CosmosPackageListV1Request: optional.NewInterface(dcosclient.CosmosPackageListV1Request{}),
-		})
-		if err != nil {
-			cErr <- err
-		} else {
-			cResult <- result
-		}
-	}()
-
-	select {
-	case resErr := <-cErr:
-		s.logger.Debug(resErr)
+	result, _, err := cosmosClient.PackageList(context.TODO(), &dcosclient.PackageListOpts{
+		CosmosPackageListV1Request: optional.NewInterface(dcosclient.CosmosPackageListV1Request{}),
+	})
+	if err != nil {
+		s.logger.Debug(err)
 		return
-	case result := <-cResult:
-		for _, pkg := range result.Packages {
-			pkgDefinition := pkg.PackageInformation.PackageDefinition
+	}
 
-			currentPkg, ok := pkgMap[pkgDefinition.Name]
-			if ok && currentPkg.ReleaseVersion >= pkgDefinition.ReleaseVersion {
-				continue
+	for _, pkg := range result.Packages {
+		pkgDefinition := pkg.PackageInformation.PackageDefinition
+
+		currentPkg, ok := pkgMap[pkgDefinition.Name]
+		if ok && currentPkg.ReleaseVersion >= pkgDefinition.ReleaseVersion {
+			continue
+		}
+		pkgMap[pkgDefinition.Name] = pkgDefinition
+	}
+
+	for _, pkg := range pkgMap {
+		wg.Add(1)
+		go func(name, version string) {
+			err := s.installPluginFromCosmos(name, version, httpClient, pbar)
+			if err != nil {
+				s.logger.Debug(err)
 			}
-			pkgMap[pkgDefinition.Name] = pkgDefinition
-		}
-
-		for _, pkg := range pkgMap {
-			wg.Add(1)
-			go func(name, version string) {
-				err := s.installPluginFromCosmos(name, version, httpClient, pbar)
-				if err != nil {
-					s.logger.Debug(err)
-				}
-				wg.Done()
-			}(pkg.Name, pkg.Version)
-		}
-	case <-time.After(3 * time.Second):
-		s.logger.Debug(errors.New("the request to get all the services installed took more than 3 seconds, not installing their CLIs"))
-		return
+			wg.Done()
+		}(pkg.Name, pkg.Version)
 	}
 }