+++ CHANGELOG.md
@@ -4,4 +4,4 @@
 
 * Features
 
-  * Auto-install CLI plugins for running services during `dcos cluster setup`
+  * Auto-install CLI plugins for running services during `dcos cluster setup` when environment variable `DCOS_CLI_EXPERIMENTAL_AUTOINSTALL_PACKAGE_CLIS` set